Update Validation-L10N-Error-Tracing.md
The old version of this contained the line
"The format for the culture name is "-", where is the language code and is the subculture code." 
The markdown was:
"The format for the culture name is "<languagecode2>-<country/regioncode2>", where <languagecode2> is the language code and <country/regioncode2> is the subculture code."
I don't know what <languagecode2> was supposed to mean, but it was displayed as an empty string. Therefore, I inserted A and B instead of <languagecode2> and <country/regioncode2> respectively.